testutil: add unique ID generation
Improve the unique IDs we were generating for logging and make
them available for all clients in this repo.

Improvements:
- Partially human-readable timestamps
- Expose timestamp extraction
- More granular, less opinionated API

Reviewed-on: https://code-review.googlesource.com/11910

+++ logging/logadmin/sinks_test.go
@@ -22,16 +22,16 @@ import (
 	"log"
 	"reflect"
 	"testing"
+	"time"
 
 	"cloud.google.com/go/internal/testutil"
-	ltesting "cloud.google.com/go/logging/internal/testing"
 	"cloud.google.com/go/storage"
 	"golang.org/x/net/context"
 	"google.golang.org/api/iterator"
 	"google.golang.org/api/option"
 )
 
-const testSinkIDPrefix = "GO-CLIENT-TEST-SINK"
+var sinkIDs = testutil.NewUIDSpace("GO-CLIENT-TEST-SINK")
 
 const testFilter = ""
 
@@ -41,8 +41,8 @@ var testSinkDestination string
 // Returns a cleanup function to be called after the tests finish.
 func initSinks(ctx context.Context) func() {
 	// Create a unique GCS bucket so concurrent tests don't interfere with each other.
-	testBucketPrefix := testProjectID + "-log-sink"
-	testBucket := ltesting.UniqueID(testBucketPrefix)
+	bucketIDs := testutil.NewUIDSpace(testProjectID + "-log-sink")
+	testBucket := bucketIDs.New()
 	testSinkDestination = "storage.googleapis.com/" + testBucket
 	var storageClient *storage.Client
 	if integrationTest {
@@ -63,12 +63,25 @@ func initSinks(ctx context.Context) func() {
 		}
 	}
 	// Clean up from aborted tests.
-	for _, sID := range ltesting.ExpiredUniqueIDs(sinkIDs(ctx), testSinkIDPrefix) {
-		client.DeleteSink(ctx, sID) // ignore error
+	it := client.Sinks(ctx)
+	for {
+		s, err := it.Next()
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			log.Printf("listing sinks: %v", err)
+			break
+		}
+		if sinkIDs.Older(s.ID, 24*time.Hour) {
+			client.DeleteSink(ctx, s.ID) // ignore error
+		}
 	}
 	if integrationTest {
-		for _, bn := range ltesting.ExpiredUniqueIDs(bucketNames(ctx, storageClient), testBucketPrefix) {
-			storageClient.Bucket(bn).Delete(ctx) // ignore error
+		for _, bn := range bucketNames(ctx, storageClient) {
+			if bucketIDs.Older(bn, 24*time.Hour) {
+				storageClient.Bucket(bn).Delete(ctx) // ignore error
+			}
 		}
 		return func() {
 			if err := storageClient.Bucket(testBucket).Delete(ctx); err != nil {
@@ -78,26 +91,6 @@ func initSinks(ctx context.Context) func() {
 		}
 	}
 	return func() {}
-}
-
-// Collect all sink IDs for the test project.
-func sinkIDs(ctx context.Context) []string {
-	var IDs []string
-	it := client.Sinks(ctx)
-loop:
-	for {
-		s, err := it.Next()
-		switch err {
-		case nil:
-			IDs = append(IDs, s.ID)
-		case iterator.Done:
-			break loop
-		default:
-			log.Printf("listing sinks: %v", err)
-			break loop
-		}
-	}
-	return IDs
 }
 
 // Collect the name of all buckets for the test project.
@@ -123,7 +116,7 @@ loop:
 func TestCreateDeleteSink(t *testing.T) {
 	ctx := context.Background()
 	sink := &Sink{
-		ID:          ltesting.UniqueID(testSinkIDPrefix),
+		ID:          sinkIDs.New(),
 		Destination: testSinkDestination,
 		Filter:      testFilter,
 	}
@@ -155,7 +148,7 @@ func TestCreateDeleteSink(t *testing.T) {
 func TestUpdateSink(t *testing.T) {
 	ctx := context.Background()
 	sink := &Sink{
-		ID:          ltesting.UniqueID(testSinkIDPrefix),
+		ID:          sinkIDs.New(),
 		Destination: testSinkDestination,
 		Filter:      testFilter,
 	}
@@ -197,7 +190,7 @@ func TestListSinks(t *testing.T) {
 	want := map[string]*Sink{}
 	for i := 0; i < 4; i++ {
 		s := &Sink{
-			ID:          ltesting.UniqueID(testSinkIDPrefix),
+			ID:          sinkIDs.New(),
 			Destination: testSinkDestination,
 			Filter:      testFilter,
 		}
